+++ src/entities/reviews/index.ts
@@ -1,0 +1,2 @@
+export { default as useReviewListQuery } from './model/hooks/useReviewListQuery';
+export { default as useReviewDetailquery } from './model/hooks/useReviewDetailQuery';
