+++ src/entities/tanstack-test/api/getTanstack.ts
@@ -1,7 +1,7 @@
 import { getApi } from '@/shared/api/axiosApis';
 import TanstackInfo from '@/entities/tanstack-test/model/types';
 
-export async function getTanstack(): Promise<TanstackInfo | null> {
+export default async function getTanstack(): Promise<TanstackInfo | null> {
   return getApi<TanstackInfo>(
     '/repos/TanStack/query',
     { test: 1 },
