+++ src/entities/tanstack-test/index.ts
@@ -1,0 +1,1 @@
+export { default as useTanstackQuery } from './model/hooks/useTanstackQuery';
