+++ src/entities/tanstack-test/model/hooks/useTanstackQuery.ts
@@ -1,8 +1,8 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { getTanstack } from '@/entities/tanstack-test/api/getTanstack';
+import getTanstack from '../../api/getTanstack';
 import TanstackInfo from '../types';
 
-export function useTanstackQuery(): UseQueryResult<TanstackInfo | null, Error> {
+export default function useTanstackQuery(): UseQueryResult<TanstackInfo | null, Error> {
   return useQuery({
     queryKey: ['tanstackRepo'],
     queryFn: getTanstack,
