+++ src/entities/tasting-list/api/getTastingList.ts
@@ -1,6 +1,6 @@
 import { getApi } from '@/shared/api/axiosApis';
 import { TastingList } from '@/entities/tasting-list/model/types';
 
-export async function getTastingList(meetingId: number): Promise<TastingList | null> {
+export default async function getTastingList(meetingId: number): Promise<TastingList | null> {
   return getApi<TastingList>(`/tasing-list/${meetingId}`);
 }
