+++ src/entities/tasting-list/index.ts
@@ -1,0 +1,1 @@
+export { default as useTastingListQuery } from './model/hooks/useTastingListQuery';
