+++ src/entities/tasting-list/model/hooks/useTastingListQuery.ts
@@ -1,8 +1,10 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { getTastingList } from '@/entities/tasting-list/api/getTastingList';
 import { TastingList } from '../types';
+import getTastingList from '../../api/getTastingList';
 
-export function useTastingListQuery(meetingId: number): UseQueryResult<TastingList | null, Error> {
+export default function useTastingListQuery(
+  meetingId: number,
+): UseQueryResult<TastingList | null, Error> {
   return useQuery({
     queryKey: ['tasting-list'],
     queryFn: () => getTastingList(meetingId),
