+++ src/features/axios-test/index.ts
@@ -1,0 +1,4 @@
+export { default as usePostTestMutation } from './model/hooks/usePostTestMutation';
+export { default as usePatchTestMutation } from './model/hooks/usePatchTestMutation';
+export { default as usePutTestMutation } from './model/hooks/usePutTestMutation';
+export { default as useDeleteTestMutation } from './model/hooks/useDeleteTestMutation';
