+++ src/features/axios-test/model/hooks/usePatchTestMutation.ts
@@ -2,6 +2,11 @@ import { useMutation, UseMutationResult } from '@tanstack/react-query';
 import { patchTest } from '../../api/patchTest';
 import { TestPost } from '../types';
 
-export function usePatchTestMutation(): UseMutationResult<TestPost | null, Error, void, unknown> {
+export default function usePatchTestMutation(): UseMutationResult<
+  TestPost | null,
+  Error,
+  void,
+  unknown
+> {
   return useMutation({ mutationFn: patchTest });
 }
