+++ src/features/axios-test/model/hooks/usePostTestMutation.ts
@@ -2,6 +2,11 @@ import { useMutation, UseMutationResult } from '@tanstack/react-query';
 import { postTest } from '../../api/postTest';
 import { TestPost } from '../types';
 
-export function usePostTestMutation(): UseMutationResult<TestPost | null, Error, void, unknown> {
+export default function usePostTestMutation(): UseMutationResult<
+  TestPost | null,
+  Error,
+  void,
+  unknown
+> {
   return useMutation({ mutationFn: postTest });
 }
