+++ src/features/login-with-kakao/index.ts
@@ -1,0 +1,1 @@
+export { default as SocialLoginButton } from './ui/KakaoLoginButton';
