+++ src/features/navigation/ui/LoginButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/shared/ui/Button';
+import { Button } from '@/shared/index';
 import Link from 'next/link';
 
 export default function LoginButton() {
