+++ src/features/navigation/ui/UserDropdown.tsx
@@ -1,5 +1,5 @@
-import { useUserStore } from '@/shared/lib/hooks/useUserStore';
-import { DropdownMenuContent, DropdownMenuGroup, DropdownMenuItem } from '@/shared/ui';
+import { useUserStore } from '@/shared/index';
+import { DropdownMenuContent, DropdownMenuGroup, DropdownMenuItem } from '@/shared/index';
 import Link from 'next/link';
 import router from 'next/router';
