--- src/features/navigation/ui/index.ts
@@ -1,5 +1,0 @@
-import LoginButton from '@/features/navigation/ui/LoginButton';
-import NavigationTabs from '@/features/navigation/ui/NavigationTabs';
-import UserDropdown from './UserDropdown';
-
-export { LoginButton, NavigationTabs, UserDropdown };
