+++ src/features/reviews/api/patchReview.ts
@@ -1,7 +1,7 @@
 import { patchApi } from '@/shared/api/axiosApis';
 import { PostReviewInfo, ReviewSucceedInfo } from '../model/types';
 
-export async function patchReview(
+export default async function patchReview(
   reviewId: number,
   post: PostReviewInfo,
 ): Promise<ReviewSucceedInfo | null> {
