+++ src/features/reviews/index.ts
@@ -1,0 +1,1 @@
+export { default as usePostReviewMutation } from './model/hooks/usePostReivewMutation';
