+++ src/features/reviews/model/hooks/usePostReivewMutation.ts
@@ -1,8 +1,8 @@
 import { useMutation, UseMutationResult } from '@tanstack/react-query';
-import { postReview } from '../../api/postReview';
 import { PostReviewInfo, ReviewSucceedInfo } from '../types';
+import postReview from '../../api/postReview';
 
-export function usePostReivewMutation(
+export default function usePostReivewMutation(
   post: PostReviewInfo,
 ): UseMutationResult<ReviewSucceedInfo | null, Error, void, unknown> {
   return useMutation({ mutationFn: () => postReview(post) });
