+++ src/shared/api/index.ts
@@ -1,0 +1,1 @@
+export { getApi, postApi, patchApi, putApi, deleteApi } from './axiosApis';
