+++ src/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './ui';
+export * from './lib/hooks';
+export * from './api';
