+++ src/shared/lib/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useUserStore } from './useUserStore';
