+++ src/shared/ui/index.ts
@@ -1,7 +1,7 @@
-import { Button } from './Button';
-import Logo from './Logo';
+export { Button } from './Button';
+export { default as Logo } from './Logo';
 
-import {
+export {
   DropdownMenu,
   DropdownMenuTrigger,
   DropdownMenuContent,
@@ -11,29 +11,12 @@ import {
   DropdownMenuSubContent,
 } from './DropdownMenu';
 
-import {
+export {
   NavigationMenu,
   NavigationMenuList,
   NavigationMenuItem,
+  NavigationMenuLink,
   NavigationMenuContent,
   NavigationMenuTrigger,
   navigationMenuTriggerStyle,
 } from './NavigationMenu';
-
-export {
-  Logo,
-  Button,
-  DropdownMenu,
-  DropdownMenuTrigger,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuSubTrigger,
-  DropdownMenuSubContent,
-  NavigationMenu,
-  NavigationMenuList,
-  NavigationMenuItem,
-  NavigationMenuContent,
-  NavigationMenuTrigger,
-  navigationMenuTriggerStyle,
-};
