+++ src/views/README.md
@@ -25,6 +25,8 @@ views /
 - **폴더** : 폴더 이름은 kebab case로 작성한다.
 - **파일** : 파일 이름은 .tsx 와 같은 컴포넌트 단위인 경우 Pascal case로 작성하고, .ts와 같은 기능 단위는 camel case로 작성한다.
 
+**각 도메인 별 페이지 view 대한 명시성을 이유로 index.ts 는 사용하지 않는다.**
+
 ## ✍️ 작성 규칙
 
 - **조립의 역할**: 이 레이어는 UI 컴포넌트를 '조립'하고 '배치'하는 역할에 집중합니다.
