+++ src/views/login/LoginPage.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { kakaoLogin } from '@/features/login-with-kakao/model/kakaoLogin';
-import SocialLoginButton from '@/features/login-with-kakao/ui/KakaoLoginButton';
+import { SocialLoginButton } from '@/features/login-with-kakao/index';
 
 export default function LoginPage() {
   return (
