+++ src/widgets/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigationBar } from './navigation-bar/NavigationBar';
